fetch/metadata/generated/video-poster-sec-fetch-site.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Site - redirection from HTTP</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(t, url) {
    var video = document.createElement('video');
    video.setAttribute('poster', url);
    document.body.appendChild(video);

    t.add_cleanup(() => video.remove());
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    induceRequest(t, makeRequestURL(key, ['sameOrigin', 'httpOrigin', 'sameOrigin']));

    return retrieve(key, { poll: true })
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
      });
  }, 'sec-fetch-site - HTTPS downgrade-upgrade');
  </script>
  </body>
</html>
